Fix 316535 Use of |signed int| instead of (unsigned) |size_t| in messages... * when SEGV trapped, report the main thread size as an unsigned size_t * Similar for memcheck overlap errors   For example, for the 2 calls:      memcpy(&a, &a, 2147483648UL);      memcpy(&a, &a, -1);  // silently accepted by gcc 4.4.4 -Wall                           // while the 3rd arg is supposed to be a size_t   we now obtain (on a 32 bit system)     Source and destination overlap in memcpy(0xbe97113f, 0xbe97113f, 2147483648)     Source and destination overlap in memcpy(0xbef6d13f, 0xbef6d13f, 4294967295)   instead of     Source and destination overlap in memcpy(0xbe8e012f, 0xbe8e012f, -2147483648)     Source and destination overlap in memcpy(0xbe8e012f, 0xbe8e012f, -1)
Do not ask me why 
   memcpy(&a, &a, -1);
is supposed to be accepted/acceptable/valid code.



git-svn-id: svn://svn.valgrind.org/valgrind/trunk@13326

+++ coregrind/m_signals.c
@@ -1716,8 +1716,8 @@ static void default_action(const vki_siginfo_t *info, ThreadId tid)
          // FIXME: assumes main ThreadId == 1
          if (VG_(is_valid_tid)(1)) {
             VG_(umsg)(
-               " The main thread stack size used in this run was %d.\n",
-               (Int)VG_(threads)[1].client_stack_szB);
+               " The main thread stack size used in this run was %lu.\n",
+               VG_(threads)[1].client_stack_szB);
          }
       }
    }
